# modules/RestBuilder/RestBuilder-impl/rest-openapi.yaml
components:
  schemas:
    Category:
      description: Category
      properties:
        name:
          description: The category name.
          type: string
        id:
          description: The category ID.
          type: integer
        tax:
          description: The tax category.
          format: double
          type: number
      type: object
    Type:
      description: Type
      properties:
        name:
          description: The type name.
          type: string
        id:
          description: The type ID.
          type: integer
        tax:
          description: the type tax.
          format: double
          type: number
      type: object
...
    Cart:
      description: Cart
      properties:
        id:
          description: The Cart ID.
          type: integer
        totalValue:
          description: total payment amount.
          format: double
          type: number
        productList:
          type: array
          items:
            type: object
            description: array de produtos
            properties:
              name:
                description: The entity name.
                type: string
              id:
                description: The entity ID.
                type: integer
              category:
                allOf:
                  - $ref: "#/components/schemas/Category"
              type:
                allOf:
                  - $ref: "#/components/schemas/Type"
              price:
                description: The price product.
                format: double
                type: number
      type: object
info:
  description: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Sale Taxe API"
  version: v1.0
openapi: 3.0.1
paths:
...
  "/cart/{cartId}":
    get:
      operationId: getCart
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
            application/xml:
              schema:
                $ref: "#/components/schemas/Cart"
          description: ""
      tags: ["Cart"]
  "cart/addProduct/carts/{cartId}/products/{productId}":
    post:
      operationId: postCartAddProductCartProduct
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: integer
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
            application/xml:
              schema:
                $ref: "#/components/schemas/Cart"
          description: ""
      tags: ["Cart"]
  "/removeProduct/carts/{cartId}/products/{productId}":
    post:
      operationId: postRemoveProductCartProduct
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: integer
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        204:
          content:
            application/json: {}
            application/xml: {}
          description: ""
      tags: ["Cart"]
  "/cart/delete/{cartId}":
    delete:
      operationId: deleteCartDeleteCart
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: integer
      responses:
        204:
          content:
            application/json: {}
            application/xml: {}
          description: ""
      tags: ["Cart"]